% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filecache.R
\name{fc.ensure_files_in_data_dir}
\alias{fc.ensure_files_in_data_dir}
\title{Ensure all given files exist in the file cache, download them if they are not.}
\usage{
fc.ensure_files_in_data_dir(pkg_info, relative_filenames, urls,
  files_are_binary = NULL, md5sums = NULL, on_errors = "warn")
}
\arguments{
\item{pkg_info, }{named list. See `fc.get_pkg_info`.}

\item{relative_filenames, }{vector of strings. A vector of filenames, realtive to the package cache.}

\item{urls, }{vector of strings. For each file, a remote URL where to download the file. Will be passed to `downloader::download`, see that function for URL encoding details.}

\item{files_are_binary, }{logical vector. For each file, whether it is binary. Only required on Windows, when files need to be downloaded. See `downloader::download` docs for details.}

\item{md5sums, }{vector of strings or NULL. A list of MD5 checksums, one for each file in param 'relative_filenames', if not NULL. If given, the files will only be reported as existing if the MD5 sums match.}

\item{on_errors, }{string. What to do if getting the files failed. One of c("warn", "stop", "ignore"). At the end, files are checked using `fc.check_files_in_data_dir`(including MD5 if given). Depending on the check results, the behaviours triggered are: "warn": Print a warning for each file that failed the check. "stop": Stop the script, i.e., the whole application. "ignore": Do nothing. You can still react using the return value.}
}
\value{
The result of the final check using `fc.check_files_in_data_dir`.
}
\description{
Ensure all given files exist in the file cache, download them if they are not.
}
