% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filecache.R
\name{fc.get_data_dir}
\alias{fc.get_data_dir}
\title{Get the absolute path of the package cache.}
\usage{
fc.get_data_dir(pkg_info)
}
\arguments{
\item{pkg_info, }{named list. Package identifier, see fc.get_pkg_info() on how to get one.}
}
\value{
string. The absolute path of the package cache. It is constructed by calling `rappdirs::user_data_dir` with the package and author names. If the author is null, the package name is also used as the author name.
}
\description{
Get the absolute path of the package cache.
}
