% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filecache.R
\name{fc.remove_local_files}
\alias{fc.remove_local_files}
\title{Delete all the given files from the package cache.}
\usage{
fc.remove_local_files(pkg_info, relative_filenames)
}
\arguments{
\item{pkg_info, }{named list. See `fc.get_pkg_info`.}

\item{relative_filenames, }{vector of strings. A vector of filenames, relative to the package cache.}
}
\value{
logical vector. For each file, whether it was deleted. Note that files which did not exist were not deleted! You should check the results using `fc.check_files_in_data_dir`.
}
\description{
Delete all the given files from the package cache.
}
