% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filecache.R
\name{files_exist_md5}
\alias{files_exist_md5}
\title{Check whether files exist, optionally with MD5 check.}
\usage{
files_exist_md5(files_absolute, md5sums = NULL)
}
\arguments{
\item{files_absolute, }{vector of strings. A vector of filenames. Files are check as given, so they must already include the package cache part of the path.}

\item{md5sums, }{vector of strings or NULL. A list of MD5 checksums, one for each file in param 'files', if not NULL. If given, the files will only be reported as existing if the MD5 sums match.}
}
\value{
logical vector. Whether the files exist. If the md5sums were given, whether the files exist and the MD5 sum matches.
}
\description{
Check whether files exist. If MD5 hashes are given, they will be verified.
}
\keyword{internal}
