% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filecache.R
\name{flatten_filepath}
\alias{flatten_filepath}
\title{Turn a filepath into a flat string.}
\usage{
flatten_filepath(filepath)
}
\arguments{
\item{filepath}{string or list of strings}
}
\value{
string, the flattened filepath
}
\description{
Turn a filepath into a flat string.
}
\keyword{internal}
