% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filecache.R
\name{get_abs_filenames}
\alias{get_abs_filenames}
\title{Join all relative filenames to a datadir.}
\usage{
get_abs_filenames(datadir, relative_filenames)
}
\arguments{
\item{datadir}{string, the path to the package cache directory.}

\item{relative_filenames, }{vector of strings. A vector of filenames, relative to the package cache.}
}
\description{
For each file, create a full path by joining the datadir with the filename.
}
\keyword{internal}
