% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filecache.R
\name{list_available}
\alias{list_available}
\title{List files that are available locally in the package cache.}
\usage{
list_available(pkg_info)
}
\arguments{
\item{pkg_info, }{named list. Package identifier, see get_pkg_info() on how to get one.}
}
\value{
vector of strings. The file names available, relative to the package cache. The returned names may include a subdirectory part. The subdirectories are not listed separately.
}
\description{
List files that are available locally in the package cache.
}
\examples{
    pkg_info = get_pkg_info("mypackage")
    available_files_in_cache = list_available(pkg_info)

}
