% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filecache.R
\name{make_pgk_cache_subdir_for_all_relative_files}
\alias{make_pgk_cache_subdir_for_all_relative_files}
\title{Given a relative file, create the subdir in the package cache if needed.}
\usage{
make_pgk_cache_subdir_for_all_relative_files(pkg_info, relative_filenames)
}
\arguments{
\item{pkg_info, }{named list. Package identifier, see get_pkg_info() on how to get one.}

\item{relative_filenames, }{vector of strings. A vector of filenames, relative to the package cache. Can be a list of vectors, which will be interpreted as files with subdirs.}
}
\description{
Given a relative file, create the subdir in the package cache if needed.
}
\keyword{internal}
