% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filecache.R
\name{make_pgk_cache_subdir_for_relative_file}
\alias{make_pgk_cache_subdir_for_relative_file}
\title{Given a relative file, create the subdir in the package cache if needed.}
\usage{
make_pgk_cache_subdir_for_relative_file(pkg_info, relative_file)
}
\arguments{
\item{pkg_info, }{named list. Package identifier, see get_pkg_info() on how to get one.}

\item{relative_file, }{string or vector of strings. If a string, this function does nothing. If a vector of strings, a path is created from the elements using file.path, and the directory of it (determined by dirname()) is created.}
}
\description{
Given a relative file, create the subdir in the package cache if needed.
}
\keyword{internal}
