% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filecache.R
\name{pkg_cache_dir_with_version}
\alias{pkg_cache_dir_with_version}
\title{Append the optional package version to a cache directory path.}
\usage{
pkg_cache_dir_with_version(dir, pkg_info)
}
\arguments{
\item{dir, }{string. The cache directory path (already including the package name).}

\item{pkg_info, }{named list. Package identifier, see get_pkg_info() on how to get one.}
}
\value{
string. The directory path, with the version appended if the package info contains one.
}
\description{
Append the optional package version to a cache directory path.
}
\keyword{internal}
