% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filecache.R
\name{fc.erase}
\alias{fc.erase}
\title{Delete the full package cache directory for the given package.}
\usage{
fc.erase(pkg_info)
}
\arguments{
\item{pkg_info, }{named list. Package identifier, see fc.get_pkg_info() on how to get one.}
}
\value{
named list. This can be passed to all function which require a `pkg_info` argument. You should not care for the inner structure and treat it as some identifier.
}
\description{
Delete the full package cache directory for the given package.
}
