% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filecache.R
\name{fc.erase}
\alias{fc.erase}
\title{Delete the full package cache directory for the given package.}
\usage{
fc.erase(pkg_info)
}
\arguments{
\item{pkg_info, }{named list. Package identifier, see fc.get_pkg_info() on how to get one.}
}
\value{
integer. The return value of the unlink() call: 0 for success, 1 for failure. See the unlink() documentation for details.
}
\description{
Delete the full package cache directory for the given package.
}
